refactor(jekyll): Abstract the model for jekyll
