--- src/jekyll/mod.rs
@@ -1,4 +1,0 @@
-pub mod jk_document;
-pub mod jk_errors;
-
-use super::error;
