+++ src/legacy_model/document.rs
@@ -8,6 +8,14 @@ impl From<DocumentBuilder> for cobalt_model::DocumentBuilder<cobalt_model::Front
     fn from(legacy: DocumentBuilder) -> Self {
         let (front, content) = legacy.parts();
         let front: cobalt_model::FrontmatterBuilder = front.into();
-        cobalt_model::DocumentBuilder::new(front, content)
+        Self::new(front, content)
+    }
+}
+
+impl From<cobalt_model::DocumentBuilder<cobalt_model::FrontmatterBuilder>> for DocumentBuilder {
+    fn from(modern: cobalt_model::DocumentBuilder<cobalt_model::FrontmatterBuilder>) -> Self {
+        let (front, content) = modern.parts();
+        let front: FrontmatterBuilder = front.into();
+        Self::new(front, content)
     }
 }
